把DC的查询移到仓储
