+++ Rafy/Plugins/Rafy.RBAC/Rafy.RBAC.DataPermissionManagement/Rafy.RBAC.DataPermissionManagement.csproj
@@ -69,7 +69,9 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="Hxy Key.snk" />
-    <None Include="packages.config" />
+    <None Include="packages.config">
+      <SubType>Designer</SubType>
+    </None>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\..\Rafy\Rafy.csproj">
