+++ Rafy/Plugins/Rafy.RBAC/Rafy.RBAC.UserRoleManagement/Rafy.RBAC.UserRoleManagement.csproj
@@ -56,6 +56,8 @@
       <DependentUpon>UserRole.cs</DependentUpon>
     </Compile>
     <Compile Include="Entities\UserRoleManagementEntity.cs" />
+    <Compile Include="Extensions\ResourceOperationRepositoryExtension.cs" />
+    <Compile Include="Extensions\ResourceRepositoryExtension.cs" />
     <Compile Include="Extensions\UserRepositoryExtension.cs" />
     <Compile Include="Extensions\RoleRepositoryExtension.cs" />
     <Compile Include="UserRoleFinder.cs" />
